+++ product_code/operation_service/impl/src/main/scala/de/upb/cs/uc4/operation/impl/OperationApplication.scala
@@ -1,6 +1,10 @@
 package de.upb.cs.uc4.operation.impl
 
+import akka.Done
 import akka.cluster.sharding.typed.scaladsl.Entity
+import akka.stream.scaladsl.Flow
+import akka.util.Timeout
+import com.lightbend.lagom.scaladsl.broker.kafka.LagomKafkaClientComponents
 import com.lightbend.lagom.scaladsl.persistence.jdbc.JdbcPersistenceComponents
 import com.lightbend.lagom.scaladsl.persistence.slick.SlickPersistenceComponents
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializerRegistry
@@ -10,24 +14,39 @@ import de.upb.cs.uc4.certificate.api.CertificateService
 import de.upb.cs.uc4.hyperledger.impl.HyperledgerComponent
 import de.upb.cs.uc4.operation.api.OperationService
 import de.upb.cs.uc4.operation.impl.actor.{ OperationDatabaseBehaviour, OperationHyperledgerBehaviour, OperationState }
+import de.upb.cs.uc4.operation.impl.commands.ClearWatchlist
 import de.upb.cs.uc4.operation.impl.readside.OperationEventProcessor
+import de.upb.cs.uc4.shared.client.JsonUsername
+import de.upb.cs.uc4.shared.client.kafka.EncryptionContainer
 import de.upb.cs.uc4.shared.server.UC4Application
+import de.upb.cs.uc4.shared.server.kafka.KafkaEncryptionComponent
+import de.upb.cs.uc4.shared.server.messages.Confirmation
+import de.upb.cs.uc4.user.api.UserService
+import de.upb.cs.uc4.user.model.JsonUserData
+import org.slf4j.{ Logger, LoggerFactory }
 import play.api.db.HikariCPComponents
+
+import scala.concurrent.Future
+import scala.concurrent.duration._
 
 abstract class OperationApplication(context: LagomApplicationContext)
   extends UC4Application(context)
   with SlickPersistenceComponents
   with JdbcPersistenceComponents
   with HikariCPComponents
-  with HyperledgerComponent {
+  with HyperledgerComponent
+  with LagomKafkaClientComponents
+  with KafkaEncryptionComponent {
 
+  private final val log: Logger = LoggerFactory.getLogger(classOf[OperationApplication])
   lazy val processor: OperationEventProcessor = wire[OperationEventProcessor]
   readSide.register(processor)
 
   override def createHyperledgerActor: OperationHyperledgerBehaviour = wire[OperationHyperledgerBehaviour]
 
-  //Bind CertificateService
+  //Bind CertificateService and UserService (for Topic)
   lazy val certificateService: CertificateService = serviceClient.implement[CertificateService]
+  lazy val userService: UserService = serviceClient.implement[UserService]
 
   // Register the JSON serializer registry
   override lazy val jsonSerializerRegistry: JsonSerializerRegistry = OperationSerializerRegistry
@@ -42,6 +61,27 @@ abstract class OperationApplication(context: LagomApplicationContext)
       entityContext => OperationDatabaseBehaviour.create(entityContext)
     )
   )
+
+  implicit val timeout: Timeout = Timeout(config.getInt("uc4.timeouts.database").milliseconds)
+
+  userService
+    .userDeletionTopicMinimal()
+    .subscribe
+    .atLeastOnce(
+      Flow.fromFunction[EncryptionContainer, Future[Done]] { container =>
+        try {
+          val jsonUsername = kafkaEncryptionUtility.decrypt[JsonUsername](container)
+          clusterSharding.entityRefFor(OperationState.typeKey, jsonUsername.username)
+            .ask[Confirmation](replyTo => ClearWatchlist(jsonUsername.username, replyTo)).map(_ => Done)
+        }
+        catch {
+          case throwable: Throwable =>
+            log.error("OperationService received invalid topic message: {}", throwable.toString)
+            Future.successful(Done)
+        }
+      }
+        .mapAsync(8)(done => done)
+    )
 }
 
 object OperationApplication {