+++ product_code/report_service/impl/src/main/scala/de/upb/cs/uc4/report/impl/actor/TextReport.scala
@@ -3,6 +3,7 @@ package de.upb.cs.uc4.report.impl.actor
 import de.upb.cs.uc4.admission.model.{ CourseAdmission, ExamAdmission }
 import de.upb.cs.uc4.certificate.model.EncryptedPrivateKey
 import de.upb.cs.uc4.course.model.Course
+import de.upb.cs.uc4.hyperledger.api.model.operation.OperationData
 import de.upb.cs.uc4.exam.model.Exam
 import de.upb.cs.uc4.examresult.model.ExamResultEntry
 import de.upb.cs.uc4.matriculation.model.ImmatriculationData
@@ -20,6 +21,8 @@ case class TextReport(
     examAdmissions: Option[Seq[ExamAdmission]],
     exams: Option[Seq[Exam]],
     examResults: Option[Seq[ExamResultEntry]],
+    operations: Seq[OperationData],
+    watchlist: Seq[String],
     timestamp: String
 )