+++ product_code/operation_service/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [WIP - v0.19.1](https://github.com/upb-uc4/University-Credits-4.0/compare/operation-v0.18.1...operation-v0.19.1) (2021-XX-XX)
+## Feature
+- Added call to get a user's watchlist
+- Added functionality to clear a user's watchlist on user deletion
+## Refactor
+## Bugfix
+
 # [v0.18.1](https://github.com/upb-uc4/University-Credits-4.0/compare/operation-v0.17.1...operation-v0.18.1) (2021-02-15)
 ## Feature
  - Adapted Service to new HLF API